backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
dnspython
django-cors-headers